+++ buildpacks/php/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-10-27
+
 ### Changed
 
 - Drop use of Classic buildpack repository (for Composer installer plugin and web server boot scripts and configs) during platform packages installation
@@ -180,7 +182,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation
 
-[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.0.10...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/heroku/buildpacks-php/compare/v1.0.10...v1.1.0
 [1.0.10]: https://github.com/heroku/buildpacks-php/compare/v1.0.9...v1.0.10
 [1.0.9]: https://github.com/heroku/buildpacks-php/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/heroku/buildpacks-php/compare/v1.0.7...v1.0.8